% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-lato.R
\docType{data}
\name{Lato}
\alias{Lato}
\alias{import_lato}
\alias{font_lato}
\alias{font_lato_bold}
\alias{font_lato_light}
\alias{font_lato_thin}
\alias{font_lato_semibold}
\alias{font_lato_heavy}
\title{Import Lato font for use in charts}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
import_lato()

font_lato

font_lato_bold

font_lato_light

font_lato_thin

font_lato_semibold

font_lato_heavy
}
\description{
Lato is Copyright (c) 2010-2014, \\u0141ukasz Dziedzic,
with Reserved Font Name Lato (\href{https://www.fontsquirrel.com/license/lato}{License}).
You can find it on \href{https://fonts.google.com/specimen/Lato}{Google Fonts}.

\code{font_lato} == "\code{Lato}" ("Regular")

\code{font_lato_bold} == "\verb{Lato Bold}"

\code{font_lato_light} == "\verb{Lato Light}"

\code{font_lato_thin} == "\verb{Lato Thin}"

\code{font_lato_semibold} == "\verb{Lato Semibold}"

\code{font_lato_heavy} == "\verb{Lato Heavy}"
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
\keyword{datasets}
