% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-source-sans-pro.R
\docType{data}
\name{SourceSansPro}
\alias{SourceSansPro}
\alias{import_source_sans}
\alias{font_ssp}
\alias{font_ssp_semibold}
\alias{font_ssp_light}
\title{Import Source Sans font for use in charts}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
import_source_sans()

font_ssp

font_ssp_semibold

font_ssp_light
}
\description{
Source Sans Pro is Copyright 2010, 2012 Adobe Systems Incorporated
(http://www.adobe.com/), with Reserved Font Name ‘Source’.
(\href{https://www.fontsquirrel.com/license/source-sans-pro}{License}).
You can find it on \href{https://fonts.google.com/specimen/Source+Sans+Pro}{Google Fonts}.

\code{font_ssp} == "\verb{Source Sans 3}" etc.
}
\details{
The font files shipped with this package were retrieved from
\verb{https://github.com/adobe-fonts/source-sans-pro/releases/download/3.028R/source-sans-3v028R.zip}.
}
\note{
Fonts are hard and platform-compatible font usage is hard to debug.
}
\keyword{datasets}
