% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tadaathemes-package.R
\docType{package}
\name{tadaathemes-package}
\alias{tadaathemes}
\alias{tadaathemes-package}
\title{tadaathemes: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tadaadata/tadaathemes}
  \item Report bugs at \url{https://github.com/tadaadata/tadaathemes/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@tadaa-data.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
