% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_readthedown.R
\name{theme_readthedown}
\alias{theme_readthedown}
\title{ggplot2 theme to fit the readthedown Rmd format}
\usage{
theme_readthedown(
  base_size = 12,
  base_family = "",
  bg = "#fcfcfc",
  axis_emph = c("xy", "x", "y"),
  ...
)
}
\arguments{
\item{base_size}{Base text size, defaults to \code{12}.}

\item{base_family}{Base text family. Use \code{"Roboto Slab"} to match the \code{readthedown}
headers, or \code{"Lato"} for the body style.}

\item{bg}{Background color, defaults to \code{rmdformats::readthedown}'s background,
\verb{#fcfcfc}.}

\item{axis_emph}{Which axis to emphasize visually (black lines).
One of \verb{"x", "y", "xy", NULL}.}

\item{...}{Other arguments passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A ggplot2 theme object.
}
\description{
A ggplot theme to fit \code{\link[rmdformats:readthedown]{rmdformats::readthedown}} in terms of background color and
dark grid lines.
}
\examples{
library(ggplot2)
p <- ggplot(iris, aes(x = Sepal.Width)) +
  geom_histogram(binwidth = .25) +
  labs(
    title = "Yet another iris plot",
    subtitle = "Using theme_readthedown",
    caption = "Oh hi there, didn't see you walk in"
  )

p + theme_readthedown()
\dontrun{
# These may require font installation
p + theme_readthedown(base_family = "Lato")
p + theme_readthedown(base_family = "Roboto Condensed", axis_emph = "x")
}
}
