% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tadaa.R
\name{theme_tadaa}
\alias{theme_tadaa}
\title{Official tadaa-data ggplot2 theme (for now)}
\usage{
theme_tadaa(
  title.size = 16,
  text.size = 14,
  legend.position = "top",
  show.axis = FALSE,
  show.grid = TRUE,
  plot.margin = c(0.8, 0.4, 0.4, 0.8)
)
}
\arguments{
\item{title.size}{title element size in pts}

\item{text.size}{text element size in pts}

\item{legend.position}{either one of "none", "left", "right", "bottom", "top"}

\item{show.axis}{boolean or "x", "y"; should axis be drawn? Which?}

\item{show.grid}{boolean; should grid lines be drawn?}

\item{plot.margin}{margin around entire plot (unit with the sizes of the top, right, bottom, and left margins)}
}
\description{
Official tadaa-data ggplot2 theme (for now)
}
