% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tadaark.R
\name{theme_tadaark}
\alias{theme_tadaark}
\title{Official dark tadaa-data ggplot2 theme (for now)}
\usage{
theme_tadaark(
  title.size = 16,
  text.size = 14,
  legend.position = "top",
  show.axis = FALSE,
  show.grid = TRUE,
  plot.margin = c(0.8, 0.4, 0.4, 0.8)
)
}
\arguments{
\item{title.size}{title element size in pts}

\item{text.size}{text element size in pts}

\item{legend.position}{either one of "none", "left", "right", "bottom", "top"}

\item{show.axis}{boolean or "x", "y"; should axis be drawn? Which?}

\item{show.grid}{boolean; should grid lines be drawn?}

\item{plot.margin}{margin around entire plot (unit with the sizes of the top, right, bottom, and left margins)}
}
\description{
Because, according to Financial Times Datavizard John Burn-Murdoch,
"they look cool".
}
\details{
Using this function changes the default color for many geoms via
\code{ggplot2::update_geom_defaults}, which, unfortunately, is rather complicated
to undo. One easy but blunt way is to restart your R session.
}
\references{
https://twitter.com/jburnmurdoch/status/1231235675229491200
}
