% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-source-sans-pro.R
\docType{data}
\name{SourceSansPro}
\alias{SourceSansPro}
\alias{import_source_sans}
\alias{font_ssp}
\alias{font_ssp_semibold}
\alias{font_ssp_bold}
\alias{font_ssp_light}
\title{Import Source Sans font for use in charts}
\format{
length 1 character vector

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
import_source_sans()

font_ssp

font_ssp_semibold

font_ssp_bold

font_ssp_light
}
\description{
Source Sans Pro is Copyright 2010, 2012 Adobe Systems Incorporated
(http://www.adobe.com/), with Reserved Font Name ‘Source’.
(\href{https://www.fontsquirrel.com/license/source-sans-pro}{License}).
You can find it on \href{https://fonts.google.com/specimen/Source+Sans+Pro}{Google Fonts}.

\code{font_ssp} == "\verb{Source Sans Pro}" etc.
}
\details{
This will add the fonts via \code{\link[sysfonts:font_add]{sysfonts::font_add}}, use
\code{\link[showtext:showtext_auto]{showtext::showtext_auto()}} to use the fonts. Other steps may be necessary.
Since this is an .otf fonts, and other fonts commonly bundled / used with R
packages (e.g. \code{hrbrthemes}) are .ttf fonts, this needs to be handled
differently, meaning that \code{\link[extrafont:font_import]{extrafont::font_import()}} will not work.
I am currently not sure which steps specifically are required for this to work,
but at least I can bundle the fonts for easier import.
}
\note{
Fonts are hard and platform-compatible font usage is hard to debug.
}
\keyword{datasets}
